+++ src/utils/subresource.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace StarkBank\Utils;
+
+
+class SubResource
+{
+    function __toArray() {
+        return get_object_vars($this);
+    }
+}
+
+?>