+++ tests/brcodePreview.php
@@ -10,7 +10,7 @@ class TestBrcodePreview
 {
     public function query()
     {
-        $previews = iterator_to_array(BrcodePreview::query(["brcodes" => ["00020126580014br.gov.bcb.pix0136a629532e-7693-4846-852d-1bbff817b5a8520400005303986540510.005802BR5908T'Challa6009Sao Paulo62090505123456304B14A"]]));
+        $previews = iterator_to_array(BrcodePreview::query(["brcodes" => ["00020126390014br.gov.bcb.pix0117valid@sandbox.com52040000530398654041.005802BR5908Jon Snow6009Sao Paulo62110507sdktest63046109"]]));
         
         if (count($previews) != 1) {
             throw new Exception("failed");